pyproject.toml
[project]
name = "n8n"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "google-api-python-client>=2.186.0",
    "google-auth>=2.42.1",
    "google-auth-oauthlib>=1.2.2",
    "line-bot-sdk>=3.20.0",
    "python-dotenv>=1.2.1",
]
